backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
pandas
numpy
websockets
aiohttp
python-dotenv
psycopg2-binary
pydantic-settings
ta
pyarrow
backtrader
matplotlib
scikit-learn
lightgbm
joblib
torch
gymnasium
mlflow
prometheus-fastapi-instrumentator
